refactor(view): extract layout/snapshot diagnostic helpers into import_validation_bridge (#3151)
Narrow verbatim extraction of pure helpers only (make_layout_rect_value /
make_layout_ancestor_chain_value / layout_runtime_snapshot_root_if_requested);
registrations, ordering, and capture_method stamping unchanged; existing
design-import + screenshot-contract suites green.

Skill-Update: skip skill=import-design reason="verbatim relocation of pure diagnostic helpers; no behavior/contract change"
